+++ AccountModelData/Account.cs
@@ -1,0 +1,9 @@
+﻿namespace AccountModelData
+{
+    public class Account
+    {
+        public int Id { get; set; }
+        public string Login { get; set; }
+        public string Password { get; set; }
+    }
+}
