--- AccountModelData/ModelData.cs
@@ -1,19 +1,0 @@
-﻿using System;
-
-namespace AccountModelData
-{
-    public class Account
-    {
-        public int Id { get; set; }
-        public string Login { get; set; }
-        public string Password { get; set; }
-    }
-    
-    public class User
-    {
-        public int Id { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
-    }
-}
