+++ AccountModelData/User.cs
@@ -1,0 +1,10 @@
+﻿namespace AccountModelData
+{
+    public class User
+    {
+        public int Id { get; set; }
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string Email { get; set; }
+    }
+}
